Refactor bot into modular structure
